+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
 - v0.19.0
-  - rename command `genautocomplete` to `autocomplete`.
+  - rename command `genautocomplete` to `autocompletion`.
   - remove command `help`.
   - change default value of option `-j` from `2` to `4`.
+  - `unikmer count/uniqs/locate`: new flag `-B/--seq-name-filter`.
 - v0.18.8
   - `unikmer info`: fix typoes.
 - v0.18.7